Refactor implementation of Image and ImageRGBA glyphs
